review fix: move code to user-utils from common
